# packages/protocol/openapi.yaml
openapi: 3.1.0
info:
  title: ClickClack API
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: http://localhost:8080
paths:
...
  /api/workspaces/{workspace_id}/projects:
    get:
      operationId: listProjects
      summary: List GitHub-linked project rooms in a workspace
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      responses:
        "200":
          description: Project list with repositories, participants, and collaboration channels
          content:
            application/json:
              schema:
                type: object
                required: [projects]
                properties:
                  projects:
                    type: array
                    items:
                      $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      summary: Create a GitHub-linked project room and collaboration channel
      parameters:
        - $ref: "#/components/parameters/workspace_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created; the webhook secret is returned only in this response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateProjectResponse"
        "403":
          description: Workspace manager permission required
...
  /api/projects/{project_id}:
    get:
      operationId: getProject
      parameters:
        - $ref: "#/components/parameters/project_id"
      responses:
        "200":
          description: Project room
          content:
            application/json:
              schema:
                type: object
                required: [project]
                properties:
                  project:
                    $ref: "#/components/schemas/Project"
  /api/projects/{project_id}/context:
    get:
      operationId: getProjectContext
      summary: Read repository, participant, and channel context for human or bot collaboration
      parameters:
        - $ref: "#/components/parameters/project_id"
      responses:
        "200":
          description: Read-only project context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectContextResponse"
...
  /api/channels/{channel_id}/messages:
    get:
      operationId: listMessages
      parameters:
        - $ref: "#/components/parameters/channel_id"
        - name: after_seq
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Root channel messages
    post:
      operationId: createMessage
      parameters:
        - $ref: "#/components/parameters/channel_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMessageRequest"
      responses:
        "200":
          description: Existing message returned for an idempotent nonce replay
        "201":
          description: Created message
...
  /api/messages/{message_id}:
    get:
      operationId: getMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      responses:
        "200":
          description: Message visible to the current user
    patch:
      operationId: updateMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "200":
          description: Updated message
    delete:
      operationId: deleteMessage
      parameters:
        - $ref: "#/components/parameters/message_id"
      responses:
        "200":
          description: Soft-deleted message
...
  /api/messages/{message_id}/thread/replies:
    post:
      operationId: createThreadReply
      parameters:
        - $ref: "#/components/parameters/message_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateThreadReplyRequest"
      responses:
        "200":
          description: Existing thread reply returned for an idempotent nonce replay
        "201":
          description: Created thread reply
...
  /api/hooks/github/projects/{project_id}:
    post:
      operationId: githubProjectWebhook
      summary: Receive signed GitHub pull request activity for a project
      security: []
      parameters:
        - $ref: "#/components/parameters/project_id"
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema:
            type: string
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
            pattern: "^sha256=[a-f0-9]{64}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: GitHub ping accepted
        "202":
          description: Delivery accepted or identified as a duplicate
        "401":
          description: Invalid project, repository, or webhook signature
        "500":
          description: Delivery processing failed
        "503":
          description: Delivery is already processing and should be retried
components:
  parameters:
    workspace_id:
      name: workspace_id
      in: path
      required: true
      schema:
        type: string
    channel_id:
      name: channel_id
      in: path
      required: true
      schema:
        type: string
    project_id:
      name: project_id
      in: path
      required: true
      schema:
        type: string
    message_id:
      name: message_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    CreateMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
        body_format:
          type: string
          enum: [markdown]
          default: markdown
        quoted_message_id:
          type: string
          description: |
            Optional ID of a message to quote. The quoted message must live in
            the same channel, DM conversation, or thread as the new message;
            cross-context quoting is rejected with 400. The server captures a
            soft snapshot of the quoted body at send time.
        nonce:
          type: string
          maxLength: 128
          description: |
            Optional client idempotency key for retry-safe message creation.
            Reusing the same nonce with the same body and quote returns the
            existing message with HTTP 200 instead of creating a duplicate.
        topic_id:
          type: string
          description: Optional topic id. Channel-scoped topics can only be used in their channel.
        kind:
          type: string
          enum: [message, agent_commentary, agent_tool]
          description: |
            Durable message kind; omitted values default to message. Agent
            activity kinds require bot-token auth
            with the explicit agent_activity:write scope and are supported on
            channel and direct-message create endpoints.
        turn_id:
          type: string
          description: |
            Optional agent-turn correlation ID. Allowed only with
            agent_commentary or agent_tool; ordinary messages reject it.
    CreateThreadReplyRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
        quoted_message_id:
          type: string
        nonce:
          type: string
          maxLength: 128
    UpdateMessageRequest:
      type: object
      required: [body]
      properties:
        body:
          type: string
...
    Channel:
      type: object
      required: [id, route_id, workspace_id, name, kind, created_at, external_managed]
      properties:
        id:
          type: string
        route_id:
          type: string
          description: Immutable public route ID used in app URLs.
        workspace_id:
          type: string
        name:
          type: string
        kind:
          type: string
        created_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
        external_managed:
          type: boolean
        external_ref:
          type: string
          description: Opaque identity in the external managing system.
        external_url:
          type: string
          format: uri
          description: Deep link into the external managing application.
        sidebar_section:
          type: string
          description: Optional client sidebar grouping label.
        last_seq:
          type: integer
          format: int64
        last_read_seq:
          type: integer
          format: int64
        unread_count:
          type: integer
          format: int64
    CreateProjectRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 80
        slug:
          type: string
        description:
          type: string
          maxLength: 500
        repositories:
          type: array
          minItems: 1
          items:
            type: string
            description: GitHub repository URL or owner/name identifier.
        github_repositories:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/GitHubAppRepositorySelection"
        member_ids:
          type: array
          items:
            type: string
    ProjectRepository:
      type: object
      required: [id, project_id, provider, owner, name, full_name, url, created_at]
      properties:
        id:
          type: string
        project_id:
          type: string
        provider:
          type: string
          enum: [github]
        owner:
          type: string
        name:
          type: string
        full_name:
          type: string
        url:
          type: string
          format: uri
        github_installation_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
    ProjectMember:
      type: object
      required: [user, role]
      properties:
        user:
          $ref: "#/components/schemas/User"
        role:
          type: string
          enum: [admin, member]
    Project:
      type: object
      required:
        [id, workspace_id, name, slug, description, channel, integration_user_id, created_by, created_at, repositories, members]
      properties:
        id:
          type: string
        workspace_id:
          type: string
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        channel:
          $ref: "#/components/schemas/Channel"
        integration_user_id:
          type: string
          description: Internal tokenless service identity used to author GitHub activity.
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        repositories:
          type: array
          items:
            $ref: "#/components/schemas/ProjectRepository"
        members:
          type: array
          items:
            $ref: "#/components/schemas/ProjectMember"
...
    GitHubAppRepositorySelection:
      type: object
      required: [installation_id, full_name]
      properties:
        installation_id:
          type: integer
          format: int64
        full_name:
          type: string
          description: Canonical GitHub owner/name identifier.
...
    CreateProjectResponse:
      type: object
      required: [project]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        webhook:
          $ref: "#/components/schemas/ProjectWebhookHandoff"
    ProjectContextResponse:
      type: object
      required: [project, context]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        context:
          type: object
          required: [source_of_truth, write_access, channel_id, repositories, participants]
          properties:
            source_of_truth:
              type: string
              enum: [github]
            write_access:
              type: boolean
              enum: [false]
            channel_id:
              type: string
            repositories:
              type: array
              items:
                $ref: "#/components/schemas/ProjectRepository"
            participants:
              type: array
              items:
                $ref: "#/components/schemas/ProjectMember"
...